Deduplicated screenshot elements
